requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0